+++ src/force/nep_small_box.cuh
@@ -190,7 +190,7 @@ static __global__ void find_descriptor_small_box(
   double* g_virial,
   float* g_sum_fxyz,
   bool need_B_projection,
-  double* B_projection,
+  float* B_projection,
   int B_projection_size)
 {
   int n1 = blockIdx.x * blockDim.x + threadIdx.x + N1;
